Add accessibility statement page and consolidate footer into shared component
- Add /accessibility statement page (WCAG 2.1 AA, mailto feedback), linked
  from the footer, modeled on MITRE's corporate accessibility language.
- Extract the duplicated footer markup from Layout.vue and index.md into a
  shared SiteFooter component (semantic <footer>), registered globally, so the
  home and doc layouts can no longer drift. Fixes the missing footer link on
  the home page.
- Ignore the local .pnpm-store/ directory.
