+++ docs/accessibility.md
@@ -1,0 +1,13 @@
+# Accessibility
+
+MITRE intends to maintain a website that is fully accessible to all individuals. We work to continually improve the user experience and apply relevant accessibility standards and guidelines.
+
+## Conformance Status
+
+The Web Content Accessibility Guidelines (WCAG) defines requirements for designers and developers to improve accessibility for people with disabilities. It defines three levels of conformance: Level A, Level AA, and Level AAA. MITRE is continuously improving the website with a goal of maintaining conformance with WCAG 2.1 AA guidelines, and WCAG 2.1 AAA whenever possible.
+
+## Provide Accessibility Feedback
+
+We welcome your feedback on the accessibility of this site. If you encounter an accessibility barrier, please let us know:
+
+[Provide Accessibility Feedback](mailto:saf@mitre.org?subject=Web%20Accessibility%20Feedback)
